+++ lib/cluster_murmur/runtime/event_dispatch_scheduler.ex
@@ -93,6 +93,7 @@ defmodule ClusterMurmur.Runtime.EventDispatchScheduler do
 
   @impl true
   def init(options) do
+    Process.flag(:trap_exit, true)
     timer_token = schedule(options.initial_delay_ms)
     {:ok, {options, %Status{cycle_count: 0, last_result: nil, last_error: nil}, timer_token}}
   end